Deleted old file (not needed anymore) + Rename file
Renames `checkAndBuild.ts` to `objectBuilder.ts` to better reflect its responsibility.

Moves the `constructFinalObject` function from `checkAndBuild.ts` to `objectBuilder.ts` and updates its import in `parseEntry.ts`.

Removes the obsolete `YINIVisitor-20250626.ts.bak` file.
