+++ src/parseEntry.ts
@@ -1,8 +1,8 @@
 import { CharStreams, CommonTokenStream } from 'antlr4'
-import { constructFinalObject } from './checkAndBuild'
 import { isDebug } from './config/env'
 import YiniLexer from './grammar/YiniLexer'
 import YiniParser, { YiniContext } from './grammar/YiniParser'
+import { constructFinalObject } from './objectBuilder'
 import { TSyntaxTreeContainer } from './types'
 import { debugPrint, printObject } from './utils/system'
 import YINIVisitor from './YINIVisitor'
